+++ integrations/batcher_test.go
@@ -61,6 +61,59 @@ func TestBatch_InvalidQuery(t *testing.T) {
 	}
 }
 
+//{"name":"packets","points":[
+// {"fields":{"value":"bad"},"time":"2015-10-18T00:00:00Z"},
+// {"fields":{"value":"good"},"time":"2015-10-18T00:00:02Z"},
+// {"fields":{"value":"good"},"time":"2015-10-18T00:00:04Z"},
+// {"fields":{"value2":"good"},"time":"2015-10-18T00:00:05Z"},
+// {"fields":{"value":"bad"},"time":"2015-10-18T00:00:06Z"},
+// {"fields":{"value":"good"},"time":"2015-10-18T00:00:08Z"}]}
+func TestBatch_ChangeDetect(t *testing.T) {
+
+	var script = `
+batch
+	|query('''
+		SELECT "value"
+		FROM "telegraf"."default".packets
+''')
+		.period(10s)
+		.every(10s)
+		.groupBy(time(2s))
+	|changeDetect('value')
+	|httpOut('TestBatch_ChangeDetect')
+`
+
+	er := models.Result{
+		Series: models.Rows{
+			{
+				Name:    "packets",
+				Tags:    nil,
+				Columns: []string{"time", "value"},
+				Values: [][]interface{}{
+					{
+						time.Date(1971, 1, 1, 0, 0, 0, 0, time.UTC),
+						"bad",
+					},
+					{
+						time.Date(1971, 1, 1, 0, 0, 2, 0, time.UTC),
+						"good",
+					},
+					{
+						time.Date(1971, 1, 1, 0, 0, 6, 0, time.UTC),
+						"bad",
+					},
+					{
+						time.Date(1971, 1, 1, 0, 0, 8, 0, time.UTC),
+						"good",
+					},
+				},
+			},
+		},
+	}
+
+	testBatcherWithOutput(t, "TestBatch_ChangeDetect", script, 21*time.Second, er, false)
+}
+
 func TestBatch_Derivative(t *testing.T) {
 
 	var script = `